LearnNepali/requirements.txt
sqlalchemy
python-dotenv
pydantic_ai
psycopg2-binary
